pyproject.toml
[project]
name = "django-cource"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.13",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "pillow>=12.2.0",
    "ruff>=0.15.11",
]
